requirements.txt
wheel
dj-database-url
dj-static
Django
python-decouple
gunicorn
psycopg2-binary
djangorestframework
drf-yasg